+++ NEWS.md
@@ -1,3 +1,20 @@
+# pctax v0.1.3 Notes
+
+## Added
+
+- added `convert_taxon_name` function <2024-11-19, Tue>
+- added `plot_two_tree` function <2024-10-22, Tue>
+
+## Fixed
+
+- fixed the file.path in windows for taxonkit <2024-11-27, Wed>
+
+# pctax v0.1.2 Notes
+
+## Added
+
+- added `topN` argument for `easy_tree()` <2024-06-25, Tue>
+
 # pctax v0.1.1 Notes
 
 ## Fixed